requirements.txt
binance-futures-connector==4.0.0
python-okx==0.2.8
celery==5.4.0
Django==5.0.4
django-celery-beat==2.6.0
django-debug-toolbar==4.3.0
flower==2.0.1
gunicorn==22.0.0
ipython==8.24.0
psycopg2-binary==2.9.9
watchfiles==0.21.0
